+++ lib/src/clixon_validate.c
@@ -1383,7 +1383,7 @@ xml_yang_validate_all(clixon_handle h,
     /* Check unique and min-max after choice test for example*/
     if (yang_config(yt) != 0){
         /* Checks if next level contains any unique list constraints */
-        if ((ret = xml_yang_minmax_recurse(xt, 1, xret)) < 0)
+        if ((ret = xml_yang_validate_minmax(xt, 1, xret)) < 0)
             goto done;
         if (ret == 0)
             goto fail;
@@ -1422,7 +1422,7 @@ xml_yang_validate_all_top(clixon_handle h,
         if ((ret = xml_yang_validate_all(h, x, xret)) < 1)
             return ret;
     }
-    if ((ret = xml_yang_minmax_recurse(xt, 0, xret)) < 1)
+    if ((ret = xml_yang_validate_minmax(xt, 0, xret)) < 1)
         return ret;
     return 1;
 }